requirements.txt
fastapi
uvicorn
sqlalchemy
streamlit
python-jose
passlib[bcrypt]
pandas
fpdf
google-generativeai
requests
